sessions/session-05-multi-agents/live-demo/wcc_tech_scout/requirements.txt
google-adk>=0.3.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
mcp>=0.1.0
